requirements.txt
python-dotenv==0.10.3
django-phonenumber-field==3.0.1
phonenumbers==8.10.21
django-localflavor==2.2
django-model-utils==3.2.0
psycopg2==2.8.3
django-crispy-forms==1.8.0
stripe==2.37.2
django-heroku==0.3.1
gunicorn==19.9.0
django-bootstrap-breadcrumbs==0.9.1
